refactor: move utils to ./lib/utils
