+++ lib/cli/build-file-anchor.js
@@ -1,5 +1,5 @@
-import escapeHtml from '../escape-html.js';
 import sanitize from '../sanitize-untrusted.js';
+import { escapeHtml } from '../utils/text.js';
 
 const SLUG_SHAPE = /^[\w.-]+\/[\w.-]+$/;
 const PATH_AND_LINE = /^(.+):(\d+)$/;
