+++ lib/cli/render-footnote.js
@@ -1,5 +1,5 @@
-import escapeHtml from '../escape-html.js';
 import { SYNTHETIC_FOOTNOTE_TEXT } from '../synthetic-footnote-text.js';
+import { escapeHtml } from '../utils/text.js';
 
 /** @import { SyntheticKey } from '../classify-message.js' */
 /** @import { ProjectResult } from './prepare-project-result.js' */
