+++ lib/cli/render-project-block.js
@@ -1,5 +1,5 @@
-import escapeHtml from '../escape-html.js';
 import sanitize from '../sanitize-untrusted.js';
+import { escapeHtml } from '../utils/text.js';
 import { renderFileSpan, renderProjectLabel } from './build-file-anchor.js';
 import { toCount } from './coerce.js';
