+++ lib/format-results.js
@@ -2,14 +2,12 @@ import chalk from 'chalk';
 import terminalLink from 'terminal-link';
 
 import aggregate from './aggregator.js';
-import escapeHtml from './escape-html.js';
-import lengthOfLongest from './length-of-longest.js';
-import pad from './pad-num.js';
 import sanitize from './sanitize-untrusted.js';
-import sortBy from './sort-by-prop.js';
-import sum from './sum-up.js';
 import { SYNTHETIC_FOOTNOTE_TEXT } from './synthetic-footnote-text.js';
 import truncateFormatterMarkdown from './truncate-formatter-markdown.js';
+import { lengthOfLongest, sortBy, sum } from './utils/array.js';
+import { pad } from './utils/format.js';
+import { escapeHtml } from './utils/text.js';
 
 const PATH_MAX = 500;
 const DEFAULT_FILE_CAP = 50;
