--- lib/pad-num.js
@@ -1,8 +1,0 @@
-/**
- * @param {number} num
- * @param {number} len
- * @returns {string}
- */
-const pad = (num, len) => num.toString().padStart(len);
-
-export default pad;
