+++ lib/sanitize-untrusted.js
@@ -1,4 +1,4 @@
-import stripControls from './strip-controls.js';
+import { stripControls } from './utils/text.js';
 
 // Secret patterns scrubbed before rendering untrusted strings into PR comment
 // markdown. An attacker-authored ESLint rule could place a token in a lint
