+++ lib/utils/array.js
@@ -1,4 +1,32 @@
 /**
+ * @template {string} T
+ * @param {T} prop
+ * @param {Array<Record<T, number>>} array
+ * @returns {number}
+ */
+function sum (prop, array) {
+  return array.reduce((count, obj) => count + obj[prop], 0);
+}
+
+/**
+ * @template {string} T
+ * @param {T} prop
+ * @param {Array<Record<T, number|string>>} input
+ * @returns {number}
+ */
+function lengthOfLongest (prop, input) {
+  let length = 0;
+
+  for (const obj of input) {
+    if (obj[prop] !== undefined) {
+      length = Math.max(length, obj[prop].toString().length);
+    }
+  }
+
+  return length;
+}
+
+/**
  * @template {Record<string, any>} T
  * @param {T} a
  * @param {T} b
@@ -31,7 +59,8 @@ const checkProp = (a, b, [prop, ...props], sortReverse) => {
  * @param {boolean} sortReverse
  * @returns {T[]}
  */
-const sortBy = (props, array, sortReverse) =>
-  array.toSorted((a, b) => checkProp(a, b, props, sortReverse));
+function sortBy (props, array, sortReverse) {
+  return array.toSorted((a, b) => checkProp(a, b, props, sortReverse));
+}
 
-export default sortBy;
+export { lengthOfLongest, sortBy, sum };
