+++ lib/utils/text.js
@@ -14,7 +14,7 @@ const CONTROLS = new Set([
  * @param {string} text
  * @returns {string}
  */
-export default function stripControls (text) {
+function stripControls (text) {
   let out = '';
   for (const ch of text) {
     const cp = ch.codePointAt(0);
@@ -22,3 +22,18 @@ export default function stripControls (text) {
   }
   return out;
 }
+
+/**
+ * @param {string} text
+ * @returns {string}
+ */
+function escapeHtml (text) {
+  return stripControls(text)
+    .replaceAll('&', '&amp;')
+    .replaceAll('<', '&lt;')
+    .replaceAll('>', '&gt;')
+    .replaceAll('"', '&quot;')
+    .replaceAll('\'', '&#39;');
+}
+
+export { escapeHtml, stripControls };
