+++ lib/write-step-summary.js
@@ -10,7 +10,7 @@ import { appendFile } from 'node:fs/promises';
  * @param {string} [callerName]  Prefix for the stderr warning. Defaults to "eslint-formatter-summary".
  * @returns {Promise<void>}
  */
-export default async function writeStepSummary (filePath, markdown, callerName = 'eslint-formatter-summary') {
+export async function writeStepSummary (filePath, markdown, callerName = 'eslint-formatter-summary') {
   try {
     // eslint-disable-next-line security/detect-non-literal-fs-filename -- caller-controlled path (typically a trusted env var)
     await appendFile(filePath, markdown + '\n', 'utf8');
